Notes issue solution w/IEquatable and Distinct
Resolved the issue where I was using Linq GroupBy after a linq join to
perform a deep search on my Work Orders Equipment where I would recieve
a "invalid memo ole or hyperlink object in subquery" when leveraging
.First() or .Single() to retrieve notes due to the notes being a Memo
field.  I resolved this by implementing IEquatable on WO in
PartialEDMXClasses.cs and then calling .Distinct() on the List<WO> in
InMemoryRepositories.cs.  This not only solved the notes issue but also
improved speed and efficiency.
I also increased the quantity of equipment that could be added to a work
order from 5 to 15