WOEquip Notes, Embedded rdlc
Native MP2 had a section for notes on equipment for a Work Order that
were stored in the WOC table (Work Order Comments). I implemented
functionality to view these in the Web App.
I also got the broad strokes for the download report to pull data into a
sub report that gets populated with all the WOEquip variables for a
given WO