Leveraged deferred execution to optomize queries
I changed the sort methods in InMemoryReqositories so that they were
included in the queries.  I was able to do this by leveraging a package
called Dynamic Linq (System.Linq.Dynamic) which allowed me to past
textual arguments to my sort methods.
I also added the CompletionDate and CompletionTime columns in favor of
the CloseDate column.  CloseDate was simply a timestamp of when the Work
Order got moved into History whereas CompletionDate and Completion time
are values that indicate when the work was actually completed.