Date Discrepancy in Reports
Removed format specifiers for both the CSV and the Excel reports as they
were causing discrepancies between MP2 dates and the reported dates for
some reason.   Also combined the Date and Time fields in the Excel
Report into a single DateTime object in the RDLC report but left them as
separate fields in the CSV report for pivot table reporting purposes.