Implemented Filter by Priority
Resolved the issue where I was unable to filter my query results by
multiple priorities.  The issue stemmed from the priority field being a
nullable double in the Access database and the SQL being generated by
Dynamic Linq to Access. I needed to convert the nullable double field to
an integer and then bring in all matches of the integer field (ie 1.1,
1.2, 1.3, etc. for when 1 was selected) that would all match the
integer(s) specified. Dynamic linq kept on creating an inappropriate IN
clause that reversed the IN clause that I wanted.  Furthermore, I had to
do quite a bit of work to get dynamic linq to access to generate the sql
to perform the conversion.
I also implemented the .skip and .take functions so that the entire
collection didn't need to get loaded (with a .ToList() call) in order
for them to properly work.  I believe that the resolution to this was in
using the Microsoft.ACE.OLEDB.12.0 provider on my connection string
instead of Microsoft.Jet.OLEDB.4.0.  Apparently Jet only really works on
32 bit servers and when I moved the application to a x64 server, I kept
getting errors thrown when a call to the database was made.