Update Initial WO Count, M WOType Added
The initial WO List showed different Work Orders than the list that MP2
generated. This was due to the fact that MP2 considered a Work Order to
be 'Opened' not by it's status, but by the fact that the 'CLOSEDATE'
field was set to 1/1/1900.  So I adjusted the initial display so that
this was also how the WO Wizard displayed Opened Work Orders.
Subsequently in order to do this I had to implement functionality for
filtering by Work Orders of type 'M' (which I am unable to manually
create) as well as Filtering by CLOSEDATE