Split Db Read and Db Write
I installed a Robocopy on the Server that monitors any changes to the
MP2 Access Database and then copies the Db to a local directory as soon
as that happens.  I then changed the code so that it READS from that
local copy of the Access Database while maintaining the WRITE to the
Network Access Database.